+++ TMessagesProj/jni/voip/webrtc/modules/audio_processing/agc2/limiter.h
@@ -17,21 +17,20 @@
 #include "modules/audio_processing/agc2/fixed_digital_level_estimator.h"
 #include "modules/audio_processing/agc2/interpolated_gain_curve.h"
 #include "modules/audio_processing/include/audio_frame_view.h"
-#include "rtc_base/constructor_magic.h"
 
 namespace webrtc {
 class ApmDataDumper;
 
 class Limiter {
  public:
-  Limiter(size_t sample_rate_hz,
+  Limiter(int sample_rate_hz,
           ApmDataDumper* apm_data_dumper,
           const std::string& histogram_name_prefix);
   Limiter(const Limiter& limiter) = delete;
   Limiter& operator=(const Limiter& limiter) = delete;
   ~Limiter();
 
-  // Applies limiter and hard-clipping to |signal|.
+  // Applies limiter and hard-clipping to `signal`.
   void Process(AudioFrameView<float> signal);
   InterpolatedGainCurve::Stats GetGainCurveStats() const;
 
@@ -40,7 +39,7 @@ class Limiter {
   // * below kMaximalNumberOfSamplesPerChannel*1000/kFrameDurationMs
   //   so that samples_per_channel fit in the
   //   per_sample_scaling_factors_ array.
-  void SetSampleRate(size_t sample_rate_hz);
+  void SetSampleRate(int sample_rate_hz);
 
   // Resets the internal state.
   void Reset();
